requirements.txt
amqp==5.1.1
asgiref==3.5.2
async-timeout==4.0.2
backports.zoneinfo==0.2.1
billiard==3.6.4.0
Brotli==1.0.9
celery==5.2.7
certifi==2022.6.15
charset-normalizer==2.1.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
ConfigArgParse==1.5.3
coreapi==2.3.3
coreschema==0.0.4
coverage==6.4.4
Deprecated==1.2.13
Django==4.1
django-environ==0.9.0
django-filter==22.1
django-log-request-id==2.0.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
drf-yasg==1.21.3
Flask==2.2.2
Flask-BasicAuth==0.2.0
Flask-Cors==3.0.10
gevent==21.12.0
geventhttpclient==2.0.2
greenlet==1.1.3
idna==3.3
importlib-metadata==4.12.0
inflection==0.5.1
itsdangerous==2.1.2
itypes==1.2.0
Jinja2==3.1.2
kombu==5.2.4
locust==2.12.1
Markdown==3.4.1
MarkupSafe==2.1.1
msgpack==1.0.4
openapi-codec==1.3.2
packaging==21.3
prompt-toolkit==3.0.31
psutil==5.9.2
psycopg2==2.9.1
PyJWT==2.4.0
pyparsing==3.0.9
pytz==2022.2.1
pyzmq==24.0.1
redis==4.3.4
requests==2.28.1
roundrobin==0.0.4
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
sentry-sdk==1.9.8
simplejson==3.17.6
six==1.16.0
sqlparse==0.4.2
typing-extensions==4.3.0
uritemplate==4.1.1
urllib3==1.26.12
vine==5.0.0
wcwidth==0.2.5
Werkzeug==2.2.2
wrapt==1.14.1
zipp==3.8.1
zope.event==4.5.0
zope.interface==5.4.0
